--- a/pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkOffheapBitmapInvertedIndexCreator.java
+++ b/pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkOffheapBitmapInvertedIndexCreator.java
@@ -1,0 +1,105 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.pinot.perf;
+
+import org.apache.commons.io.FileUtils;
+import org.apache.pinot.core.segment.creator.impl.inv.OffHeapBitmapInvertedIndexCreator;
+import org.apache.pinot.spi.data.DimensionFieldSpec;
+import org.apache.pinot.spi.data.FieldSpec;
+import org.openjdk.jmh.annotations.*;
+import org.openjdk.jmh.runner.Runner;
+import org.openjdk.jmh.runner.options.ChainedOptionsBuilder;
+import org.openjdk.jmh.runner.options.OptionsBuilder;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+@State(Scope.Benchmark)
+public class BenchmarkOffheapBitmapInvertedIndexCreator {
+
+  public enum Assignment {
+    ROUND_ROBIN {
+      @Override
+      void assign(OffHeapBitmapInvertedIndexCreator creator, int docs, int cardinality) {
+        for (int i = 0; i < docs; ++i) {
+          creator.add(i % cardinality);
+        }
+      }
+    },
+    SORTED_UNIFORM {
+      @Override
+      void assign(OffHeapBitmapInvertedIndexCreator creator, int docs, int cardinality) {
+        for (int i = 0; i < cardinality; ++i) {
+          for (int j = 0; j < docs / cardinality; ++j) {
+            creator.add(i);
+          }
+        }
+      }
+    };
+
+    abstract void assign(OffHeapBitmapInvertedIndexCreator creator, int docs, int cardinality);
+  }
+
+  private Path indexDir;
+  @Param({"10", "1000", "10000"})
+  int cardinality;
+
+  @Param({"1000000", "10000000", "100000000"})
+  int numDocs;
+
+  @Param
+  Assignment assignment;
+
+  private OffHeapBitmapInvertedIndexCreator creator;
+
+  @Setup(Level.Invocation)
+  public void setup() throws IOException {
+    indexDir = Files.createTempDirectory("index");
+    creator = new OffHeapBitmapInvertedIndexCreator(
+            indexDir.toFile(), new DimensionFieldSpec("foo", FieldSpec.DataType.STRING, true),
+            cardinality, numDocs, -1);
+    assignment.assign(creator, numDocs, cardinality);
+  }
+
+  @TearDown(Level.Invocation)
+  public void tearDown() throws IOException {
+    if (null != indexDir) {
+      FileUtils.deleteDirectory(indexDir.toFile());
+    }
+    creator.close();
+  }
+
+  @Benchmark
+  public Object seal() throws IOException {
+    creator.seal();
+    return creator;
+  }
+
+  public static void main(String[] args)
+          throws Exception {
+    ChainedOptionsBuilder opt =
+            new OptionsBuilder().include(BenchmarkOffheapBitmapInvertedIndexCreator.class.getSimpleName())
+                    .mode(Mode.SingleShotTime)
+                    .warmupIterations(8).measurementIterations(8).forks(5);
+
+    new Runner(opt.build()).run();
+  }
+}