@richardstartin 2020-12-04
Regarding the test failure, this is because the bitmap is being run length encoded incrementally. It becomes deterministic when setting

```java
_bitmapWriters[i] = RoaringBitmapWriter.bufferWriter().runCompress(false).get();
```

To make the test deterministic, RLE could be disabled, but this is bad for file size. Or the off-heap index creator could run-optimise its bitmaps, but that's bad for speed compared to doing it incrementally. Alternatively the test could be modified to check that the bitmaps are equal after memory mapping.