@richardstartin 2020-12-05
The solution I came up with to make sure the same bytes are produced by both on-heap and off-heap creators was to apply the same changes to `OffHeapBitmapInvertedIndexCreator`, which will now also be run-optimised incrementally.