@siddharthteotia 2020-12-05
Can you follow one of the perf test classes (module pinot-perf , e.g BenchmarkPinotDataBuffer, BenchmarkDictionary etc) and get some numbers?

Also, you can write a simple test function that can be called while simultaneously profiling with YourKit/VisualVM. This can help reveal how the cpu and memory usage changes with this code.