@kishoreg 2020-12-05
This looks promising both from performance and memory allocation. @Jackie-Jiang can you please look into this?