@Jackie-Jiang 2020-12-05
Find this blog from @richardstartin explaining the optimization very well: https://richardstartin.github.io/posts/roaringbitmaps-construction-performance